Fix aggregator tests. Remove aggregate_xxx.go files. Add blobInner/blobOuter migrations and refactor proof table
