+++ aggregator/blobouter.go
@@ -5,3 +5,7 @@ import "context"
 func (a *Aggregator) tryGenerateBlobOuterProof(ctx context.Context, prover proverInterface) (bool, error) {
 	return false, nil
 }
+
+func (a *Aggregator) tryAggregateBlobOuterProofs(ctx context.Context, prover proverInterface) (bool, error) {
+	return false, nil
+}
